+++ concordium-sdk/src/main/java/com/concordium/sdk/ClientV2.java
@@ -5,17 +5,18 @@ import com.concordium.grpc.v2.Empty;
 import com.concordium.grpc.v2.QueriesGrpc;
 import com.concordium.sdk.exceptions.BlockNotFoundException;
 import com.concordium.sdk.exceptions.ClientInitializationException;
-import com.concordium.sdk.exceptions.TransactionNotFoundException;
 import com.concordium.sdk.requests.BlockHashInput;
 import com.concordium.sdk.requests.getaccountinfo.AccountRequest;
 import com.concordium.sdk.responses.BlockIdentifier;
 import com.concordium.sdk.responses.accountinfo.AccountInfo;
 import com.concordium.sdk.responses.blockinfo.BlockInfo;
 import com.concordium.sdk.responses.blocksummary.updates.queues.AnonymityRevokerInfo;
+import com.concordium.sdk.responses.blocksummary.updates.queues.IdentityProviderInfo;
 import com.concordium.sdk.responses.consensusstatus.ConsensusStatus;
 import com.concordium.sdk.responses.cryptographicparameters.CryptographicParameters;
 import com.concordium.sdk.responses.transactionstatus.TransactionStatus;
 import com.concordium.sdk.transactions.AccountAddress;
+import com.concordium.sdk.transactions.AccountTransaction;
 import com.concordium.sdk.transactions.AccountNonce;
 import com.concordium.sdk.transactions.Transaction;
 import com.concordium.sdk.transactions.BlockItem;
@@ -172,6 +173,31 @@ public final class ClientV2 {
     }
 
     /**
+     * Sends an Account Transaction to the Concordium Node.
+     *
+     * @param accountTransaction Account Transaction to send.
+     * @return Transaction {@link Hash}.
+     */
+    public Hash sendTransaction(final AccountTransaction accountTransaction) {
+        var req = ClientV2MapperExtensions.to(accountTransaction);
+        var grpcOutput = this.server().sendBlockItem(req);
+
+        return to(grpcOutput);
+    }
+
+    /**
+     * Gets all the Identity Providers at the end of the block pointed by {@link BlockHashInput}.
+     *
+     * @param input Pointer to the Block.
+     * @return {@link Iterator} of {@link IdentityProviderInfo}
+     */
+    public Iterator<IdentityProviderInfo> getIdentityProviders(final BlockHashInput input) {
+        var grpcOutput = this.server().getIdentityProviders(to(input));
+
+        return to(grpcOutput, ClientV2MapperExtensions::to);
+    }
+
+    /**
      * Get the {@link CryptographicParameters} at a given block.
      *
      * @param blockHash the hash of the block
@@ -188,7 +214,7 @@ public final class ClientV2 {
             throw BlockNotFoundException.from(blockHash.getBlockHash());
         }
     }
-    
+
     /**
      * Retrieves a {@link BlockInfo}
      *